CodeGen: Move LibcallLoweringInfo to separate file
